apps/backend/requirements.txt
gTTS>=2.3.0
edge-tts
faster-whisper
g2p_en
nltk
numpy


piper-tts
